+++ Controller/ProcessController.php
@@ -20,6 +20,7 @@ use Kontrolgruppen\CoreBundle\Filter\ProcessFilterType;
 use Kontrolgruppen\CoreBundle\Form\ProcessCompleteType;
 use Kontrolgruppen\CoreBundle\Form\ProcessType;
 use Kontrolgruppen\CoreBundle\Repository\ProcessRepository;
+use Kontrolgruppen\CoreBundle\Repository\ProcessStatusRepository;
 use Kontrolgruppen\CoreBundle\Repository\ServiceRepository;
 use Kontrolgruppen\CoreBundle\Repository\UserRepository;
 use Kontrolgruppen\CoreBundle\Service\LogManager;
@@ -318,16 +319,17 @@ class ProcessController extends BaseController
     /**
      * @Route("/{id}/complete", name="process_complete", methods={"GET","POST"})
      *
-     * @param \Symfony\Component\HttpFoundation\Request               $request
-     * @param \Kontrolgruppen\CoreBundle\Entity\Process               $process
-     * @param \Kontrolgruppen\CoreBundle\Repository\ServiceRepository $serviceRepository
+     * @param \Symfony\Component\HttpFoundation\Request                     $request
+     * @param \Kontrolgruppen\CoreBundle\Entity\Process                     $process
+     * @param \Kontrolgruppen\CoreBundle\Repository\ServiceRepository       $serviceRepository
+     * @param \Kontrolgruppen\CoreBundle\Repository\ProcessStatusRepository $processStatusRepository
      *
      * @return \Symfony\Component\HttpFoundation\Response
      *
      * @throws \Doctrine\ORM\NoResultException
      * @throws \Doctrine\ORM\NonUniqueResultException
      */
-    public function complete(Request $request, Process $process, ServiceRepository $serviceRepository): Response
+    public function complete(Request $request, Process $process, ServiceRepository $serviceRepository, ProcessStatusRepository $processStatusRepository): Response
     {
         $this->denyAccessUnlessGranted('edit', $process);
 
@@ -338,7 +340,11 @@ class ProcessController extends BaseController
             );
         }
 
-        $form = $this->createForm(ProcessCompleteType::class, $process);
+        $availableStatuses = $processStatusRepository->getAvailableCompletingStatusForProcessType($process->getProcessType());
+
+        $form = $this->createForm(ProcessCompleteType::class, $process, [
+            'available_statuses' => $availableStatuses,
+        ]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {