+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-[KON-317](https://jira.itkdev.dk/browse/KON-317): Added possibility for changing net value for completed processes by service
+- [KON-317](https://jira.itkdev.dk/browse/KON-317): Added possibility for changing net value for completed processes by service
+- [KON-330](https://jira.itkdev.dk/browse/KON-330): Mark statuses for use when completing processes
+
+### Fixed
+- [KON-371](https://jira.itkdev.dk/browse/KON-371): Show available statuses on process based on process type
 
 ## [1.6.7] - 2020-05-25
 ### Fixed